GEMSE: Expose Representative computation as 'difference marker' filter

+++ modules/GEMSe/iARepresentative.cpp
@@ -25,7 +25,7 @@
 #include <QVector>
 
 template <class T>
-void diff_marker_tmpl(QVector<iAITKIO::ImagePointer> imgsBase, int differenceMarkerValue, iAITKIO::ImagePointer & result)
+void diff_marker_tmpl(QVector<iAITKIO::ImagePointer> imgsBase, double differenceMarkerValue, iAITKIO::ImagePointer & result)
 {
 	typedef itk::Image<T, iAITKIO::m_DIM > ImgType;
 	QVector<ImgType*> imgs;
@@ -58,7 +58,7 @@ void diff_marker_tmpl(QVector<iAITKIO::ImagePointer> imgsBase, int differenceMar
 	result = out;
 }
 
-iAITKIO::ImagePointer CalculateDifferenceMarkers(QVector<iAITKIO::ImagePointer> imgs, int differenceMarkerValue)
+iAITKIO::ImagePointer CalculateDifferenceMarkers(QVector<iAITKIO::ImagePointer> imgs, double differenceMarkerValue)
 {
 	if (imgs.size() == 0) // all child images filtered out
 	{
